#164-Board클래스의 writerId 타입 수정.

+++ app/src/main/java/com/teamdonut/eatto/data/Board.java
@@ -35,8 +35,7 @@ public class Board {
 
     @SerializedName("writer_id")
     @Expose
-    private int writerId;
-
+    private long writerId;
 
     @SerializedName("write_date")
     @Expose
